feat: Refactor document handling and user state management; streamline data saving process
